build: Save binary with debugging info in Distribution/oscam-XXXX.debug
Before to get a binary with debugging info you have to rebuild Oscam
with DEBUG=1 parameter set to `make` because by default the outputed
binary was stripped.

This created additional work when debugging and developing because
you have to remember to run different command for building and also
it was making it harder for our users when there was crash and we
needed a backtrace.

This commit eases the pain. Now along with the stripped binary
that is saved in Distribution/ we keep unstripped binary with the
same name and .debug prefix. That way if there is problem that needs
debugging we can just ask the user to use oscam-XXX.debug binary
instead of asking for rebuilds.