Rewrite ecm whitelist code.
Change the structure that tracked ecm whitelist and rewrite the
related code. Now it is a lot simpler (160 lines less code).
There is a positive effect on binary size as well (-1.2k).

bloatcheck report:
{{{
 add/remove: 4/2 grow/shrink: 0/5 up/down: 1059/-2357 (-1298)
 function                                     old     new   delta
 chk_ecm_whitelist                              -     533    +533
 mk_t_ecm_whitelist                             -     425    +425
 ecm_whitelist_add                              -      67     +67
 clear_ecm_whitelist                            -      34     +34
 static.ecm_ratelimit_check                  2745    2728     -17
 matching_reader                             2043    1973     -70
 static.free_reader_ecm_whitelist             149       -    -149
 static.ecm_ratelimit_findspace              2685    2499    -186
 update_ecmlen_from_stat                      709     340    -369
 mk_t_ecmwhitelist                            463       -    -463
 ecmwhitelist_fn                             1232     129   -1103
}}}

A negative is that after the rewrite the ecm whitelist structure
takes a bit more memory than before.

This fixes small memory leak in parsing shown in trac
ticket #4099 comment 11.