fixed buffer overflows when sending/receiving ecms/emms
the rules for ecms are now:
min. size: 0 or >=3 (a size of 1 or 2 is not allowed)
max. size: MAX_ECM_SIZE (defined in globals.h)
SCT_LEN(ecm) must be <= er->ecmlen

the rules for emms are now:
min. size: 0 (we could add also a check for >=3 here, but I am not sure atm whether this would affect emm reassembly)
max. size: MAX_EMM_SIZE (defined in globals.h)